Revisions of October 2017
Extensive revisions of the template and numerous revisions of the class file.
In short, the changes implemented the formatting requested by OGS office
in my August meeting with Sharon, Jade and Kathy.  All of the requested
formatting requirements have been implemented and the text content of the
template has been revised to be consistent with their terminology and
desired illustrations/examples/demos.